fix timer and answer
0.0000 timer to 0 timer
check correct answer

+++ Assets/Script/EndController.cs
@@ -5,13 +5,12 @@ using UnityEngine.UI;
 
 public class EndController : MonoBehaviour
 {
-
     [SerializeField] Text clearTimeText;
     [SerializeField] Text pointText;
     // Start is called before the first frame update
     void Start()
     {
-        clearTimeText.text = GameManager.Instance.clearTime.ToString() + " Sec";
+        clearTimeText.text = ((int)GameManager.Instance.clearTime).ToString() + " Sec";
         pointText.text = GameManager.Instance.point.ToString() + " Point";
     }
